point-uv-diffusion/requirements.txt
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorlog==6.9.0
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
dash==3.0.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
einops==0.8.1
embreex==2.17.7.post6
exceptiongroup==1.2.2
executing==2.2.0
fake-bpy-module==20250325
fastjsonschema==2.21.1
filelock==3.13.1
Flask==3.0.3
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.6.1
ftfy==6.3.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
hydra-core==1.3.2
idna==3.4
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.34.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lazy_loader==0.4
lightning-utilities==0.14.2
lxml==5.3.1
manifold3d==3.0.1
mapbox_earcut==1.0.3
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mistune==3.1.3
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath==1.3.0
multidict==6.2.0
narwhals==1.32.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.3
notebook==7.3.3
notebook_shim==0.2.4
numpy==1.24.4
nvdiffrast @ file:///vsc-hard-mounts/leuven-data/375/vsc37593/my_py_projects/nvdiffrast
nvidia-ml-py==12.570.86
nvitop==1.4.2
omegaconf==2.3.0
open3d==0.18.0
opencv-python-headless==4.6.0.66
overrides==7.7.0
packaging==24.2
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.7
plotly==6.0.1
point-cloud-utils==0.31.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycollada==0.9
pycparser==2.22
pyg-lib==0.4.0+pt21cu121
Pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.3
pyquaternion==0.9.9
pyrootutils==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
pytorch-lightning==2.2.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.23.1
rtree==1.4.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
Send2Trash==1.8.3
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
svg.path==6.3
sympy==1.13.1
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.3.30
timm==1.0.15
tinycss2==1.4.0
tomli==2.2.1
torch==2.1.0+cu121
torch-cluster==1.6.3+pt21cu121
torch-geometric==2.6.1
torch-scatter==2.1.2+pt21cu121
torch-sparse==0.6.18+pt21cu121
torch-spline-conv==1.2.2+pt21cu121
torchaudio==2.1.0+cu121
torchinfo==1.8.0
torchmetrics==1.7.0
torchsummary==1.5.1
torchvision==0.16.0+cu121
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.6.5
triton==2.1.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
urllib3==1.26.13
vhacdx==0.0.8.post2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.13
xatlas==0.0.9
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
